refactor(validation): polish metadata plugin internals
