+++ packages/validation/src/utils/findJsonPointerNode.ts
@@ -1,0 +1,23 @@
+import {Document, isCollection, ParsedNode} from 'yaml';
+
+export function findJsonPointerNode(valuesDoc: Document.Parsed<ParsedNode>, path: string[]) {
+  if (!valuesDoc.contents) {
+    return undefined;
+  }
+
+  let valueNode: any = valuesDoc.contents;
+
+  for (let c = 0; valueNode && c < path.length; c += 1) {
+    let node = path[c];
+    if (isCollection(valueNode)) {
+      const nextNode = valueNode.get(node, true);
+      if (nextNode) {
+        valueNode = nextNode;
+      } else {
+        return valueNode;
+      }
+    } else break;
+  }
+
+  return valueNode;
+}
